create generic

+++ libs/libregister/libregister.h
@@ -1,8 +1,8 @@
 #ifndef _LIBREGISTER_H_
 #define _LIBREGISTER_H_
 
-#include "../../include/thread_struct.h"
-#include "../../include/register.h"
+#include "../../generic/include/thread_struct.h"
+#include "../../generic/include/register.h"
 
 #define FILEPATH    "/dev/hapara_reg"
 
@@ -10,6 +10,6 @@ void libregister_test(void);
 void print_struct(struct hapara_thread_struct *sp);
 
 int reg_add(struct hapara_thread_struct *thread_info);
-int reg_del(unsigned int off, uint8_t target);
+int reg_del(unsigned int off, unsigned int target);
 
 #endif
